+++ packages/target-tauri/src-tauri/build.rs
@@ -71,6 +71,9 @@ fn main() {
             "html_email_open_menu",
             "html_email_set_load_remote_content",
             "update_tray_icon_badge",
+            "get_available_themes",
+            "load_theme",
+            "get_current_active_theme",
         ]),
     ))
     .expect("failed to run tauri-build");